Customize serialization routine of Histogram (wip)
The motivation is that this alleviates the dependency on std::vector
serialization, which allows the use of some drop-in replacements of
std::vector, such as folly::fbvector.

Note that currently it is still not really POD, since make_array() is
still used and the size of data is not fixed. And there is a memory leak
due to the missing delete [].